pyproject.toml
[project]
name = "cloud-automation-with-python"
version = "0.1.0"
description = "Lecture task solutions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"boto3>=1.37.0",
	"python-dotenv>=1.0.1",
	"python-magic>=0.4.27",
]

[tool.uv]
package = false
